Amendments 17-27
I describe what the amendments say, how they have been interpreted, their histories, and my opinions. I link to sources in text when they are only relevant for one sentence or one part of a sentence, or at the end of a series of sentences if they were relevant for multiple.