Made fake data for mockup
File contains multiple named "tables" (just arrays) with fake data, mimicking the agreed-upon DB schema. Demonstrates that can have items of the same name but in different locations; low stock of 10 ohm resistors in Mercer Lab means we can show a warning to user 2, who is an admin of that location; reservations include suspicious records (a student taking 50 of an item at once and a student not returning a borrowable item), which the admin could see and then take action on.