Admin Homepage Markup; Preliminary JavaScript
I wrote the markup for the admin homepage, so it has all the elements shown in the wireframe, but it isn't styled so looks ugly for now.
I added an (empty) external stylesheet and JavaScript file in a resources folder. They should hold stuff that will be used on multiple pages (e.g. styling for the navbar, the onClick function for the menu button that will be in the top right).
I wrote some JavaScript functions in resources/script.js that are currently called to populate the admin homepage, but can be called with any userID. They only work if the info from data.json has been put in localStorage as a string value with the key "data", so you should do that in a callback from $(document).ready(), as demonstrated in the <head> of admin_homepage.html. They also require jQuery, you can link to a CDN, also seen in admin_homepage.html.