Log-in/Sign-up markup & styling; CSS separation
Now, CSS that applies to multiple pages in in resources/style.css, and all other CSS is either in the page's <head> or inline. I wrote the markup & styling for the login and sign up pages in line with the wireframes. They are so far non-functional, I still need to write the JS.