Admin homepage JS fix
Now all functions that need it are passed the JSON data as an argument. The JSON is used to be put in local storage and then each function got it out of local storage when it was called, which was a silly way to do things and not very optimized.