Expanded general homepage
Expanded the general homepage by creating a guideline for adding more information to the page.

The guideline is as follows:
Each blurb of information should contain the following:
A title for what it is about displayed on the left hand side
An image showcasing what the blurb is talking about
A description on the bottom giving details related to the subject.
Optionally, the blurb may contain a button with a link.