User & Admin unification
Now both user and admin homepages use the userid cookie to figure out which data to show. If that cookie or the admin cookie is not set, or if they are a user on the admin page or an admin on the user page, it redirects back to the login page.
The admin homepage now has the same header and styling as the user homepage.
When a user successfully logs in, it deletes the old values of their userid and admin cookies.