moved menu dropdown code
Since the menu dropdown is standard throughout the whole site once the user logs in, I filled the original "menuClick()" function in script.js to show the dropdown of the menu.

This keeps the dropdown code to one page for easy editing later on if new links need to added since we only need to edit the menu.