removed unnecessary jQuery; some code refactoring
Now the only jQuery used is in making AJAX calls to fetch the JSON, which is unavoidable but will be eliminated when we write the backend (we'll just query the DB). I also eliminated some repetitive code, and moved some code out of script.js and into a new file in resources, lablist.js, since those functions are only used by the admin and user homepages so we shouldn't waste time loading them for other pages that don't need them.