Accessibility improvements
Increased the color contrast in some places, and fixed a JS bug that was logging errors to the console.

There are still accessibility problems wrt color contrast. The blueish-gray color we use throughout the site (#778DA9) doesn't contrast enough with either black OR white for Lighthouse to be consistently happy, but there are times when it doesn't mind so I'm not sure.